# Umniah-E-Sport.Tool/OpenAPIs/swagger5.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Zain-E-Sport.API",
    "version": "v1"
  },
  "paths": {
...
    "/api/UserGame/get-users-by-game": {
      "get": {
        "tags": [
          "UserGame"
        ],
        "parameters": [
          {
            "name": "GameId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetUsersByGameVM"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetUsersByGameVM"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetUsersByGameVM"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/UserGame/get-users-games": {
      "get": {
        "tags": [
          "UserGame"
        ],
        "parameters": [
          {
            "name": "UserId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "GameId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetUsersGamesVM"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUsersGamesVM"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUsersGamesVM"
                }
              }
            }
          }
        }
      }
    },
    "/api/UserGame/get-user-games-by-userId": {
      "get": {
        "tags": [
          "UserGame"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetUserGamesByUserIdVm"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetUserGamesByUserIdVm"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetUserGamesByUserIdVm"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Achievement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "name_EN": {
            "type": "string",
            "nullable": true
          },
          "name_AR": {
            "type": "string",
            "nullable": true
          },
          "imageName": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Arena": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name_EN": {
            "type": "string",
            "nullable": true
          },
          "name_AR": {
            "type": "string",
            "nullable": true
          },
          "countryName_EN": {
            "type": "string",
            "nullable": true
          },
          "countryName_AR": {
            "type": "string",
            "nullable": true
          },
          "imageName": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "tournaments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tournament"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Game": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title_EN": {
            "type": "string",
            "nullable": true
          },
          "title_AR": {
            "type": "string",
            "nullable": true
          },
          "imageName": {
            "type": "string",
            "nullable": true
          },
          "shortCode": {
            "type": "string",
            "nullable": true
          },
          "gameLink": {
            "type": "string",
            "nullable": true
          },
          "termsAndConditions": {
            "$ref": "#/components/schemas/TermsAndConditions"
          },
          "termsAndConditionsId": {
            "type": "integer",
            "format": "int32"
          },
          "tournaments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tournament"
            },
            "nullable": true
          },
          "videos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Video"
            },
            "nullable": true
          },
          "userGames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserGames"
            },
            "nullable": true
          },
          "news": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewsEntity"
            },
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "GetUserGamesByUserIdVm": {
        "type": "object",
        "properties": {
          "gameId": {
            "type": "integer",
            "format": "int32"
          },
          "userScore": {
            "type": "integer",
            "format": "int32"
          },
          "title_EN": {
            "type": "string",
            "nullable": true
          },
          "title_AR": {
            "type": "string",
            "nullable": true
          },
          "gameImageName": {
            "type": "string",
            "nullable": true
          },
          "gameShortCode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GetUsersByGameVM": {
        "type": "object",
        "properties": {
          "gameId": {
            "type": "integer",
            "format": "int32"
          },
          "userScore": {
            "type": "integer",
            "format": "int32"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "GetUsersGamesVM": {
        "type": "object",
        "properties": {
          "gameId": {
            "type": "integer",
            "format": "int32"
          },
          "userScore": {
            "type": "integer",
            "format": "int32"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "Location": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "text_EN": {
            "type": "string",
            "nullable": true
          },
          "text_AR": {
            "type": "string",
            "nullable": true
          },
          "tournaments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tournament"
            },
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "NewsEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "game": {
            "$ref": "#/components/schemas/Game"
          },
          "gameId": {
            "type": "integer",
            "format": "int32"
          },
          "imageName": {
            "type": "string",
            "nullable": true
          },
          "title_EN": {
            "type": "string",
            "nullable": true
          },
          "title_AR": {
            "type": "string",
            "nullable": true
          },
          "description_EN": {
            "type": "string",
            "nullable": true
          },
          "description_AR": {
            "type": "string",
            "nullable": true
          },
          "timeStamp": {
            "type": "string",
            "format": "date-time"
          },
          "creationTimeStamp": {
            "type": "string",
            "format": "date-time"
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "TermsAndConditions": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "gameId": {
            "type": "integer",
            "format": "int32"
          },
          "game": {
            "$ref": "#/components/schemas/Game"
          },
          "content_EN": {
            "type": "string",
            "nullable": true
          },
          "content_AR": {
            "type": "string",
            "nullable": true
          },
          "endTimeStamp": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "creationTimeStamp": {
            "type": "string",
            "format": "date-time"
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "TimeSpan": {
        "type": "object",
        "properties": {
          "ticks": {
            "type": "integer",
            "format": "int64"
          },
          "days": {
            "type": "integer",
            "format": "int32"
          },
          "hours": {
            "type": "integer",
            "format": "int32"
          },
          "milliseconds": {
            "type": "integer",
            "format": "int32"
          },
          "minutes": {
            "type": "integer",
            "format": "int32"
          },
          "seconds": {
            "type": "integer",
            "format": "int32"
          },
          "totalDays": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "totalHours": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "totalMilliseconds": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "totalMinutes": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "totalSeconds": {
            "type": "number",
            "format": "double",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "Tournament": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "gameId": {
            "type": "integer",
            "format": "int32"
          },
          "game": {
            "$ref": "#/components/schemas/Game"
          },
          "arena": {
            "$ref": "#/components/schemas/Arena"
          },
          "arenaId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "title_EN": {
            "type": "string",
            "nullable": true
          },
          "title_AR": {
            "type": "string",
            "nullable": true
          },
          "description_EN": {
            "type": "string",
            "nullable": true
          },
          "description_AR": {
            "type": "string",
            "nullable": true
          },
          "imageName": {
            "type": "string",
            "nullable": true
          },
          "bannerName": {
            "type": "string",
            "nullable": true
          },
          "organizer_EN": {
            "type": "string",
            "nullable": true
          },
          "organizer_AR": {
            "type": "string",
            "nullable": true
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "tournamentType": {
            "$ref": "#/components/schemas/TournamentType"
          },
          "tournamentTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "locationId": {
            "type": "integer",
            "format": "int32"
          },
          "prizeAmount": {
            "type": "integer",
            "format": "int32"
          },
          "isFeatured": {
            "type": "boolean"
          },
          "subscriptionFee": {
            "type": "integer",
            "format": "int32"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "creationTimeStamp": {
            "type": "string",
            "format": "date-time"
          },
          "durationTimeSpan": {
            "$ref": "#/components/schemas/TimeSpan"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "discordLink": {
            "type": "string",
            "nullable": true
          },
          "userTournaments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserTournaments"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TournamentType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "text_EN": {
            "type": "string",
            "nullable": true
          },
          "text_AR": {
            "type": "string",
            "nullable": true
          },
          "tournaments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tournament"
            },
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "msisdn": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "creationTimeStamp": {
            "type": "string",
            "format": "date-time"
          },
          "unsubscribeTimeStamp": {
            "type": "string",
            "format": "date-time"
          },
          "subscribeTimeStamp": {
            "type": "string",
            "format": "date-time"
          },
          "isSubscribe": {
            "type": "boolean"
          },
          "userTournaments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserTournaments"
            },
            "nullable": true
          },
          "achievements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Achievement"
            },
            "nullable": true
          },
          "userGames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserGames"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserGames": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "game": {
            "$ref": "#/components/schemas/Game"
          },
          "gameId": {
            "type": "integer",
            "format": "int32"
          },
          "userScore": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "UserTournaments": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "ingameId": {
            "type": "string",
            "nullable": true
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "tournament": {
            "$ref": "#/components/schemas/Tournament"
          },
          "tournamentId": {
            "type": "integer",
            "format": "int32"
          },
          "timeStamp": {
            "type": "string",
            "format": "date-time"
          },
          "isTheFirst": {
            "type": "boolean"
          },
          "isJoinSmsSent": {
            "type": "boolean"
          },
          "isDailySmsSent": {
            "type": "boolean"
          },
          "isHourlySmsSent": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Video": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "game": {
            "$ref": "#/components/schemas/Game"
          },
          "gameId": {
            "type": "integer",
            "format": "int32"
          },
          "imageName": {
            "type": "string",
            "nullable": true
          },
          "videoFileName": {
            "type": "string",
            "nullable": true
          },
          "title_EN": {
            "type": "string",
            "nullable": true
          },
          "title_AR": {
            "type": "string",
            "nullable": true
          },
          "description_EN": {
            "type": "string",
            "nullable": true
          },
          "description_AR": {
            "type": "string",
            "nullable": true
          },
          "timeStamp": {
            "type": "string",
            "format": "date-time"
          },
          "creationTimeStamp": {
            "type": "string",
            "format": "date-time"
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    }
  }
}